+++ src/Geoapify.SDK/Geocoding/Inputs/Filters/Filter.cs
@@ -4,8 +4,8 @@ public abstract class Filter
 {
 	internal abstract string ToQueryString();
 
-	public static CountryFilter ByCountry(params IEnumerable<CountryCode> countryCodes)
+	public static CountryCodeFilter ByCountryCodes(params ICollection<CountryCode> countryCodes)
 	{
-		return new CountryFilter(countryCodes);
+		return new CountryCodeFilter(countryCodes);
 	}
 }